Fix env template names

+++ starfish/install.sh
@@ -2,13 +2,16 @@
 set -e
 set -x
 
-cp .env.template .env 
+if [ ! -f ./.env ]; then
+    cp .env.template .env 
+    read -p "You will need to edit .env to add your secret keys. Press Enter to confirm you understand and continue."
+fi
 
-read -p "You will need to edit .env to add your secret keys. Press Enter to confirm you understand and continue."
 
 python3 -m venv .venv
 source .venv/bin/activate
 
+source ./.env
 pip install -r requirements.txt
 pip install -r requirements_dev.txt
